P2 rebase

+++ hal/src/nRF52840/exflash_hal.c
@@ -323,7 +323,7 @@ static int erase_common(uintptr_t start_addr, size_t num_blocks, nrf_qspi_erase_
     /* Round down to the nearest block */
     start_addr = ((start_addr / block_length) * block_length);
 
-    for (int i = 0; i < num_blocks; i++) {
+    for (size_t i = 0; i < num_blocks; i++) {
         err_code = nrfx_qspi_erase(len, start_addr);
         if (err_code) {
             goto erase_common_done;
